Overhaul design to improve CLASP and SWASP throughput.
The main daemon has been split into a management
daemon (pipelined) and a worker daemon for each
camera (pipeline_workerd). This allows the
processing to occur on a different machine
per-camera, to support the new SuperWASP design.

Furthermore, each worker daemon runs several
processes which allow frames to be processed
in parallel. Synchronisation to ensure that
previews and guide profiles are updated in
monotonic order is managed through shared
memory based on the DATE-OBS  timestamps.

Each of the worker daemon processes spawns
a thread that is used to update ds9 previews
without blocking the main processing flow.